requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
streamlit>=1.40.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
